feat: Bundler plugin built on `unplugin`
