+++ packages/unplugin-typegpu/package.json
@@ -57,10 +57,11 @@
     "prepare-package": "tgpu-dev-cli prepack"
   },
   "dependencies": {
-    "tinyest-for-wgsl": "workspace:~0.1.0-alpha.1",
+    "@babel/standalone": "^7.26.6",
     "estree-walker": "^3.0.3",
     "magic-string": "^0.30.11",
-    "@babel/standalone": "^7.26.6"
+    "tinyest-for-wgsl": "workspace:~0.1.0-alpha.1",
+    "unplugin": "^2.2.0"
   },
   "devDependencies": {
     "@babel/template": "^7.25.9",
