+++ packages/unplugin-typegpu/src/esbuild.ts
@@ -1,0 +1,3 @@
+import { esbuildPlugin } from './index';
+
+export default esbuildPlugin;
