+++ packages/unplugin-typegpu/src/farm.ts
@@ -1,0 +1,3 @@
+import { farmPlugin } from './index';
+
+export default farmPlugin;
