+++ packages/unplugin-typegpu/src/rolldown.ts
@@ -1,0 +1,3 @@
+import { rolldownPlugin } from './index';
+
+export default rolldownPlugin;
