+++ packages/unplugin-typegpu/src/rspack.ts
@@ -1,0 +1,3 @@
+import { rspackPlugin } from './index';
+
+export default rspackPlugin;
