+++ packages/unplugin-typegpu/src/vite.ts
@@ -1,0 +1,3 @@
+import { vitePlugin } from './index';
+
+export default vitePlugin;
