+++ packages/unplugin-typegpu/src/webpack.ts
@@ -1,0 +1,3 @@
+import { webpackPlugin } from './index';
+
+export default webpackPlugin;
